compiling qual/non-qual xdrop
